// src/app/models/auditoriaNivelItRequisito.ts
export class AuditoriaNivelItRequisito {
  id: BigInteger;
  auditoria_nivel_item_id: BigInteger;
  auditoria_requisito_id: BigInteger;
  ds_requisito: string;
  classificacao_requisito_id: BigInteger;
  nr_peso: number;
  nm_classificacao: string;
  ds_orientacao: string;
  ie_evidencia: string;
}